105 | Fixing docstrings by including resources correctly
The docstrings were not built for some modules as a consequence of parmeterising variables to config/parameters.yaml as
I mistakenly loaded this file using an absolute path. When building the documentation in the "doc" directory that file
could not be found.

The correct way to do this is to include the files as resources and that has now been implemented using pkgutil which
loads the file from pgfinder/config/parameters.yaml as a byte string which is then read with yaml.safe_load().

Useful reading...

https://stackoverflow.com/a/58941536 (and associated examples in https://github.com/wimglenn/resources-example)
https://stackoverflow.com/a/51724506 (single clear example of the pkgutil that is include in above solutions)