Restore column order and round theo_mwMonoisotopic
The `theo_mwMonoisotopic` has been rounded to 4 decimal places now, but
I'm not sure if this is the right thing to do? Perhaps to make things
consistent we should round *all* of the columns?

This column was rounded in this commit because floating-point error made
it print something silly like: 724.312700000000001 — I thought that was
rather ugly?