Hack to fix version in regression tests
`setuptools_scm` sets the version based on the tag and git commit in relation to it. This means that now that the
version is included in the metadata that tests are no longer reproducible because each commit changes the version.
This is solved by fixing the version recorded in the target test files for regression tests and during the tests to be
`version: pgfinder` and so this doesn't impact the results of the regression tests.