Switching to RTD theme; API autobuild fixed
In an attempt to get multiple versions of docs built on ReadTheDocs I've switched themes to the RTD theme. At the same
time I noticed the automatic building of API documentation was not happening so I've (hopefully) rendered that and it
will be available across all versions going forward.

Not yet sure whether the "[flyout](https://docs.readthedocs.io/en/stable/flyout-menu.html)" menu will be rendered, its
not shown locally and a [thread](https://stackoverflow.com/questions/66436744/how-to-activate-the-rtd-flyout-menu)
referenced problems in the past and there are some issues on GitHub but these appear to have been resolved.