Compare dataframes via Pandas testing framework
Switches to using Pandas Testing framework to compare data frames.

In doing this I've noticed that there are, as noted by @TheLostLambda, duplicate column names in the consolidated
results file.

This is generally a bad idea as...

a) it can lead to confusion for people working with the files.
b) its awkward having dataframes (or any data structure for that matter) with non-unique names and makes subsequent
   development awkward.

Not dealing with this now will only come back to bite us in the future, and it is noted in the unit test that was
introduced in 2d4543f as causing problems already.

To this end I have introduced a 'suffix' argument to the `consolidate_result()` which by default is `candidate` and
makes columns from `consolidate_result()` unique.

The suffix is appended (within parentheses) to the consolidated columns, although if a column already has
`(consolidated)` in its name (as is the case for `Inferred structure (consolidated)` and `Itensity (consolidated)`) then
the suffix is included within the existing parentheses (i.e. `(consolidated candidate)`)