Minimum Python version bumped to 3.9
Tidying up doc strings we are switching to the newer format of `str | Path` rather than the older `Union[str,
Path]`. This can be supported under Python 3.8 with `from __future__ import annotations` but since Python 3.8 is due to
reach End of Life in [2024-10-31](https://endoflife.date/python) there is no harm in dropping support now and setting
the minimum version to 3.9 and adding support/testing for 3.11.