Parametrises the column names of input and derived/output files
+ Documentation on output columns added.
+ Parametrises columns in `lib/pgfinder/config/columns.yaml` and loads these on initialisation.
+ Uses the columns defined and loaded in `pgfinder.find_pg` functions for loading ftrs and maxquant files.
+ Adds a dictionary for mapping and renaming consolidated variables.

Abstracting the information out of hard-coded variables into parameter files simplifies the process of extending to
future update file formats and makes managing changes (still to come in this PR) easier.

Will have to add to documentation how the configuration file works and perhaps also something to change it, although the
way things are going I don't think anyone will use this interactively and so this documentation would be for
developers (which doesn't make it any less important as the more we can help our future selves the better).