tests(web): Narrow elements in search_custom_database_test
With the introduction of new elements in the WebUI the existing
`web/tests/search_custom_database_test.ts` was failing as the wrong elements were selected (the new added ones rather
than the existing ones).

This has been corrected by adding `data-testid` fields to the Svelte pages <div> sections that define the specific areas
to give them unique identifiers and updating the tests to use `page.getByTestID()` targets to align with these.